move existing signedpod's circuits draft to its own file
