+++ src/backends/plonky2/circuits/mod.rs
@@ -1,2 +1,3 @@
 pub mod common;
 pub mod mainpod;
+pub mod signedpod;
